Move CSSProcessor to its own namespace
